move tools for restructuring apps, roles, etc from core to admin
