+++ houdini/houdini_admin/apps.py
@@ -1,5 +1,5 @@
 from django.apps import AppConfig
 
 
-class AdminConfig(AppConfig):
-    name = 'admin'
+class HoudiniAdminConfig(AppConfig):
+    name = 'houdini_admin'
