+++ houdini/houdini_admin/templates/houdini_admin/401.html
@@ -1,4 +1,4 @@
-{% extends 'houdini_server/base.html' %}
+{% extends 'houdini_admin/base.html' %}
 {% block content %}
 
 <h1>401 Unauthorized</h1>
