+++ houdini/houdini_admin/templates/houdini_admin/applications.html
@@ -1,4 +1,4 @@
-{% extends 'houdini_server/base.html' %}
+{% extends 'houdini_admin/base.html' %}
 {% block content %}
 <a class="btn btn-success" href="">Create Application</a>
 <table class="table table-striped">
