+++ houdini/houdini_admin/templates/houdini_admin/base.html
@@ -6,7 +6,7 @@
 		<title>Houdini</title>
 		<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css" integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u" crossorigin="anonymous">
 		<link rel="stylesheet" type="text/css" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css">
-		<link rel="stylesheet" type="text/css" href="{% static 'houdini_server/css/style.css' %}">
+		<link rel="stylesheet" type="text/css" href="{% static 'houdini_admin/css/style.css' %}">
 		<script src="http://code.jquery.com/jquery-2.2.4.min.js"   integrity="sha256-BbhdlvQf/xTY9gja0Dq3HiwQF8LaCRTXxZKRutelT44="   crossorigin="anonymous"></script>
 		<script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js" integrity="sha384-Tc5IQib027qvyjSMfHjOMaLkfuWVxZxUPnCJA7l2mCWNIpG9mGCD8wGNIcPD7Txa" crossorigin="anonymous"></script>
 	</head>
@@ -34,7 +34,7 @@
 			</nav>
 			<div class="content container-fluid">
 				<div class="row">
-			        {% include 'houdini_server/messages.html' %}
+			        {% include 'houdini_admin/messages.html' %}
 					{% block content %}
 					{% endblock %}
 				</div>
