+++ houdini/houdini_admin/templates/houdini_admin/create.html
@@ -1,4 +1,4 @@
-{% extends 'houdini_server/base.html' %}
+{% extends 'houdini_admin/base.html' %}
 {% block content %}
 <div class="col-md-4 col-md-offset-4">
     {% if object %}
