+++ houdini/houdini_admin/templates/houdini_admin/directories.html
@@ -1,4 +1,4 @@
-{% extends 'houdini_server/base.html' %}
+{% extends 'houdini_admin/base.html' %}
 {% block content %}
 <a class="btn btn-success" href="/directories/create">Create Directory</a>
 <table class="table table-striped">
