+++ houdini/houdini_admin/templates/houdini_admin/index.html
@@ -1,0 +1,1 @@
+{% extends 'houdini_admin/base.html' %}
