+++ houdini/houdini_admin/templates/houdini_admin/permission_test.html
@@ -1,4 +1,4 @@
-{% extends 'houdini_server/base.html' %}
+{% extends 'houdini_admin/base.html' %}
 {% block content %}
 
 <h3>Permission Test</h3>
