+++ houdini/houdini_admin/templates/houdini_admin/permissions.html
@@ -1,4 +1,4 @@
-{% extends 'houdini_server/base.html' %}
+{% extends 'houdini_admin/base.html' %}
 {% block content %}
 <div class="col-md-6 col-md-offset-3">
     <h2>
