+++ houdini/houdini_admin/templates/houdini_admin/roles.html
@@ -1,4 +1,4 @@
-{% extends 'houdini_server/base.html' %}
+{% extends 'houdini_admin/base.html' %}
 {% block content %}
 <a class="btn btn-success">Create Role</a>
 <table class="table table-striped">
