+++ houdini/houdini_admin/templates/houdini_admin/table.html
@@ -1,4 +1,4 @@
-{% extends 'houdini_server/base.html' %}
+{% extends 'houdini_admin/base.html' %}
 {% block content %}
 <div class="col-md-10 col-md-offset-1">
     {% if messages %}
