+++ houdini/houdini_admin/templates/houdini_admin/users.html
@@ -1,0 +1,3 @@
+{% extends 'houdini_admin/base.html' %}
+{% block content %}
+{% endblock %}
