+++ houdini/houdini_client/templates/houdini_client/register.html
@@ -1,4 +1,4 @@
-{% extends 'houdini_server/base.html' %}
+{% extends 'houdini_admin/base.html' %}
 {% block content %}
 
 <h3>Register</h3>
