--- houdini/houdini_server/templates/houdini_server/index.html
@@ -1,1 +1,0 @@
-{% extends 'houdini_server/base.html' %}
