--- houdini/houdini_server/templates/houdini_server/users.html
@@ -1,3 +1,0 @@
-{% extends 'houdini_server/base.html' %}
-{% block content %}
-{% endblock %}
